+++ src/main/java/roomescape/common/config/WebConfig.java
@@ -2,14 +2,10 @@ package roomescape.common.config;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestClient;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.common.exception.handler.PaymentExceptionHandler;
-import roomescape.infrastructure.payment.toss.TossPaymentWithRestClient;
 import roomescape.presentation.interceptor.AdminAuthInterceptor;
 import roomescape.presentation.resvoler.LoginMemberArgumentResolver;
