+++ src/main/java/roomescape/infrastructure/payment/toss/exception/PaymentExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.common.exception.handler;
+package roomescape.infrastructure.payment.toss.exception;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -11,8 +11,6 @@ import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.web.client.ResponseErrorHandler;
 import roomescape.common.exception.PaymentClientException;
 import roomescape.common.exception.PaymentServerException;
-import roomescape.infrastructure.payment.toss.TossPaymentErrorCodeForClient;
-import roomescape.infrastructure.payment.toss.TossPaymentErrorCodeForServer;
 
 public class PaymentExceptionHandler implements ResponseErrorHandler {
 
@@ -23,43 +21,43 @@ public class PaymentExceptionHandler implements ResponseErrorHandler {
 
     @Override
     public void handleError(URI url, HttpMethod method, ClientHttpResponse response)
-        throws IOException {
+            throws IOException {
         InputStream inputStream = response.getBody();
         TossPaymentErrorResponse tossPaymentErrorResponse = parseToTossPaymentErrorResponse(
-            inputStream);
+                inputStream);
         checkClientError(tossPaymentErrorResponse);
         checkServerError(tossPaymentErrorResponse);
     }
 
     private static void checkServerError(TossPaymentErrorResponse tossPaymentErrorResponse) {
         Optional<TossPaymentErrorCodeForServer> tossPaymentErrorCodeForServer = TossPaymentErrorCodeForServer.of(
-            tossPaymentErrorResponse.code);
+                tossPaymentErrorResponse.code);
         if (tossPaymentErrorCodeForServer.isPresent()) {
             throw new PaymentServerException(
-                tossPaymentErrorCodeForServer.get().getMessage());
+                    tossPaymentErrorCodeForServer.get().getMessage());
         }
     }
 
     private static void checkClientError(TossPaymentErrorResponse tossPaymentErrorResponse) {
         Optional<TossPaymentErrorCodeForClient> tossPaymentErrorCodeForClient = TossPaymentErrorCodeForClient.of(
-            tossPaymentErrorResponse.code);
+                tossPaymentErrorResponse.code);
         if (tossPaymentErrorCodeForClient.isPresent()) {
             throw new PaymentClientException(
-                tossPaymentErrorCodeForClient.get().getMessage());
+                    tossPaymentErrorCodeForClient.get().getMessage());
         }
     }
 
     private TossPaymentErrorResponse parseToTossPaymentErrorResponse(InputStream bodyStream)
-        throws IOException {
+            throws IOException {
         ObjectMapper mapper = new ObjectMapper().configure(
-            DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES,
-            false);
+                DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES,
+                false);
         return mapper.readValue(bodyStream, TossPaymentErrorResponse.class);
     }
 
     private record TossPaymentErrorResponse(
-        String message,
-        String code
+            String message,
+            String code
     ) {
 
     }
